requirements.txt
Django==4.1.4
psycopg2-binary==2.9.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
celery==4.4.1
redis==3.4.1
django-redis==5.2.0
black==22.12.0
isort==5.11.4
factory-boy==3.2.1
pytest==7.2.0
pytest-django==4.5.2
mock==5.0.0
autoflake==2.0.0
requests
html5lib
bs4